Update the container to Ubuntu 18.04 and nginx 1.14.0
This updates the nginx-src builder to use Ubuntu 18.04 (as base
container) and nginx 1.14.0 (latest stable). I've not yet switched to
dynamic modules, since part of the fun is that it will be a single
binary you can copy to your systems.

I've switched to a patched version of ngx_http_upstream_fair because of
changes that landed in nginx 1.11 (and thus it was no longer compiling).
The original repo was https://github.com/gnosek/nginx-upstream-fair/,
the new one is https://github.com/itoffshore/nginx-upstream-fair/.

The README is updated (and corrected) and a more extensive usage example
(with a multi-stage build) has been added as well.